+++ src/main/java/com/example/soopgwan/domain/achive/domain/AchieveSuccess.java
@@ -1,0 +1,41 @@
+package com.example.soopgwan.domain.achive.domain;
+
+import com.example.soopgwan.domain.user.domain.User;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.Id;
+import javax.persistence.IdClass;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import java.time.LocalDate;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@Builder
+@IdClass(AchieveSuccessId.class)
+@Table(name = "tbl_achieve_success")
+@Entity
+public class AchieveSuccess {
+
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "achieve_id", nullable = false)
+    private Achieve achieve;
+
+    @Column(columnDefinition = "DATE", nullable = false)
+    private LocalDate date;
+}
