+++ src/main/java/com/example/soopgwan/domain/achive/domain/AchieveSuccessId.java
@@ -1,0 +1,19 @@
+package com.example.soopgwan.domain.achive.domain;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@EqualsAndHashCode
+public class AchieveSuccessId implements Serializable {
+
+    private Long user;
+
+    private Long achieve;
+}
