+++ src/main/java/com/example/soopgwan/domain/habit/domain/HabitSuccess.java
@@ -6,34 +6,36 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.Access;
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.IdClass;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-import java.time.LocalDate;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Builder
-@Table(name = "tbl_habit")
+@IdClass(HabitSuccessId.class)
+@Table(name = "tbl_achive_success")
 @Entity
-public class Habit {
+public class HabitSuccess {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(columnDefinition = "VARCHAR(200)", nullable = false)
-    private String content;
+    @Id
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "habit_id", nullable = false)
+    private Habit habit;
 
-    @Column(columnDefinition = "DATE", nullable = false)
-    private LocalDate startAt;
-
-    @Column(columnDefinition = "DATE", nullable = false)
-    private LocalDate endAt;
-
-    private Long count;
+    @Column(columnDefinition = "BOOLEAN", nullable = false)
+    private Boolean success;
 }
