+++ src/main/java/com/example/soopgwan/domain/habit/domain/HabitSuccessId.java
@@ -1,0 +1,19 @@
+package com.example.soopgwan.domain.habit.domain;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@EqualsAndHashCode
+public class HabitSuccessId implements Serializable {
+
+    private Long id;
+
+    private Long habit;
+}
